+++ extension/chrome/elements/compose-modules/compose-recipients-module.ts
@@ -11,8 +11,8 @@ import { Str, Value } from '../../../js/common/core/common.js';
 import { ApiErr } from '../../../js/common/api/error/api-error.js';
 import { Bm, BrowserMsg } from '../../../js/common/browser/browser-msg.js';
 import { Catch } from '../../../js/common/platform/catch.js';
-import { Google } from '../../../js/common/api/google.js';
-import { GoogleAuth } from '../../../js/common/api/google-auth.js';
+import { Google } from '../../../js/common/api/email-provider/gmail/google.js';
+import { GoogleAuth } from '../../../js/common/api/email-provider/gmail/google-auth.js';
 import { Lang } from '../../../js/common/lang.js';
 import { Ui } from '../../../js/common/browser/ui.js';
 import { Xss } from '../../../js/common/platform/xss.js';
