+++ extension/chrome/settings/modules/backup-automatic-module.ts
@@ -10,7 +10,7 @@ import { UnreportableError } from '../../../js/common/platform/catch.js';
 import { Ui } from '../../../js/common/browser/ui.js';
 import { ApiErr } from '../../../js/common/api/error/api-error.js';
 import { Assert } from '../../../js/common/assert.js';
-import { GoogleAuth } from '../../../js/common/api/google-auth.js';
+import { GoogleAuth } from '../../../js/common/api/email-provider/gmail/google-auth.js';
 import { KeyStore } from '../../../js/common/platform/store/key-store.js';
 import { KeyUtil } from '../../../js/common/core/crypto/key.js';
