+++ extension/chrome/settings/modules/experimental.ts
@@ -8,7 +8,7 @@ import { Browser } from '../../../js/common/browser/browser.js';
 import { BrowserMsg } from '../../../js/common/browser/browser-msg.js';
 import { Buf } from '../../../js/common/core/buf.js';
 import { Catch } from '../../../js/common/platform/catch.js';
-import { GoogleAuth } from '../../../js/common/api/google-auth.js';
+import { GoogleAuth } from '../../../js/common/api/email-provider/gmail/google-auth.js';
 import { Lang } from '../../../js/common/lang.js';
 import { Settings } from '../../../js/common/settings.js';
 import { Ui } from '../../../js/common/browser/ui.js';
