+++ extension/js/common/api/email-provider/gmail/gmail.ts
@@ -16,8 +16,8 @@ import { Catch } from '../../../platform/catch.js';
 import { Contact, KeyUtil } from '../../../core/crypto/key.js';
 import { Env } from '../../../browser/env.js';
 import { FormatError } from '../../../core/crypto/pgp/msg-util.js';
-import { Google } from '../../google.js';
-import { GoogleAuth } from '../../google-auth.js';
+import { Google } from './google.js';
+import { GoogleAuth } from './google-auth.js';
 import { Mime } from '../../../core/mime.js';
 import { PgpArmor } from '../../../core/crypto/pgp/pgp-armor.js';
 import { SendableMsg } from '../sendable-msg.js';
