+++ extension/js/common/api/email-provider/gmail/google-auth.ts
@@ -5,19 +5,19 @@
 // tslint:disable:no-direct-ajax
 // tslint:disable:oneliner-object-literal
 
-import { GOOGLE_API_HOST, GOOGLE_OAUTH_SCREEN_HOST, FLAVOR } from '../core/const.js';
-import { Url, Value } from '../core/common.js';
-import { tabsQuery, windowsCreate } from '../browser/chrome.js';
-import { Api } from './api.js';
-import { ApiErr } from './error/api-error.js';
-import { Backend } from './backend.js';
-import { Buf } from '../core/buf.js';
-import { Catch } from '../platform/catch.js';
-import { GmailRes } from './email-provider/gmail/gmail-parser';
-import { GoogleAuthErr } from './error/api-error.js';
-import { GoogleAuthWindowResult$result } from '../browser/browser-msg.js';
-import { Ui } from '../browser/ui.js';
-import { AcctStore, AcctStoreDict } from '../platform/store/acct-store.js';
+import { GOOGLE_API_HOST, GOOGLE_OAUTH_SCREEN_HOST, FLAVOR } from '../../../core/const.js';
+import { Url, Value } from '../../../core/common.js';
+import { tabsQuery, windowsCreate } from '../../../browser/chrome.js';
+import { Api } from './../../api.js';
+import { ApiErr } from './../../error/api-error.js';
+import { Backend } from './../../backend.js';
+import { Buf } from '../../../core/buf.js';
+import { Catch } from '../../../platform/catch.js';
+import { GmailRes } from './gmail-parser';
+import { GoogleAuthErr } from './../../error/api-error.js';
+import { GoogleAuthWindowResult$result } from '../../../browser/browser-msg.js';
+import { Ui } from '../../../browser/ui.js';
+import { AcctStore, AcctStoreDict } from '../../../platform/store/acct-store.js';
 
 type GoogleAuthTokenInfo = { issued_to: string, audience: string, scope: string, expires_in: number, access_type: 'offline' };
 type GoogleAuthTokensResponse = { access_token: string, expires_in: number, refresh_token?: string, id_token: string, token_type: 'Bearer' };
