+++ extension/js/common/api/email-provider/gmail/google.ts
@@ -4,13 +4,13 @@
 
 // tslint:disable:no-direct-ajax
 
-import { Api, ProgressCbs, ReqMethod } from './api.js';
-import { Dict, Str } from '../core/common.js';
+import { Api, ProgressCbs, ReqMethod } from '../../api.js';
+import { Dict, Str } from '../../../core/common.js';
 
-import { GOOGLE_API_HOST } from '../core/const.js';
-import { GmailRes } from './email-provider/gmail/gmail-parser.js';
+import { GOOGLE_API_HOST } from '../../../core/const.js';
+import { GmailRes } from './gmail-parser.js';
 import { GoogleAuth } from './google-auth.js';
-import { Serializable } from '../platform/store/abstract-store.js';
+import { Serializable } from '../../../platform/store/abstract-store.js';
 
 export class Google {
